Flexability and performance
If lists are set to not editable, we don't want to load all attributes because it's faster to only load what we see in the list (especially avoiding children). However we may have calculated attributes that need certain attributes to work.

Add the ability to flag which attributes must "always load" in order for business logic in the client not to break if absent but necessary.

Add checks on feather to not bother with property logic if we're not loading properties in a list.