Fix sort on composite type
Update view creation to find natural key of child object and make sure it's first attribute in child view so a sort on the child view will sort on the natural key. Takes advantage of an inherit postgres behavior.

Also make "getFeathers" a standard function on feather service since this is the third time we've needed it.