Dialog visibility check
Make sure if we leave a page with a dialog open, that it is still in an open state when we come back.

Use case is we want to search on a relation widget, or create a new value for a relation widget from within a dialog. We'll leave the page where the dialog is, then comeback with the relation widget populated.