Add resourceLink.tableData and blob formatter utility
Add tableData definition for resourceLink, which paints based on whether an icon or label were specified.

A utility to convert data: urls to blob is included for when a resource is provided as a data: url, the link is generated with a blob reference versus trying to mash the entire data url into the link.